ref (back-models): Separation des modèles pour plus de clarté
