+++ backend/gimao/models/__init__.py
@@ -1,0 +1,20 @@
+# Import des modèles
+from .avoir import Avoir
+from .consommable import Consommable
+from .constituer import Constituer
+from .correspondre import Correspondre
+from .defaillance import Defaillance
+from .document_defaillance import DocumentDefaillance
+from .document_intervention import DocumentIntervention
+from .document_technique import DocumentTechnique
+from .equipement import Equipement
+from .est_compatible import EstCompatible
+from .fabricant import Fabricant
+from .fournisseur import Fournisseur
+from .information_statut import InformationStatut
+from .intervention import Intervention
+from .lieu import Lieu
+from .modele_equipement import ModeleEquipement
+from .role import Role
+from .stock_consommable import StockConsommable
+from .validators import *
