+++ backend/gimao/models/role.py
@@ -1,0 +1,7 @@
+from django.db import models
+
+class Role(models.Model):
+    nomRole = models.CharField(max_length=50)
+
+    def __str__(self):
+        return self.nomRole
